Correct image in deployment examples
The image for the merlin daemon was incorrectly named in the example
deployment files.

Further we set the log level to the default value, as an empty string
(""), are causing issue, and perhaps it can be confusing that you either
have to no set it at all, or set it to a valid value.

In docker-compose the containers have been renamed to match the images.

Part of: MON-12873

Signed-off-by: Jacob Hansen <jhansen@op5.com>